+++ src/py/cb.plot_codeword_hist.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python3.7
+#!/usr/bin/env python3
 # coding=utf-8
 import pandas as pd
 import matplotlib.pyplot as plt